notes-backend/requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
django-environ==0.9.0